pkg/fileutils: move pattern-matching code to separate files
